Enable downloading photo and meta data
We are able to download the photo and meta data by downloadd.py.
Before using it, please check if you set everything in config.ini.
config.ini.SAMPLE is an example of config.ini. Because config.ini
contains the password of your face2gene account, please don't commit
it in the repository.

Another requirement for downloading is the folder for all the JSON files
in PEDIA cohort. We use it to define the case list in our cohort.
We will fix it in the near future.